requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
Django==1.8.6
django-bootstrap3==6.2.2
ecdsa==0.13
Fabric==1.10.2
gunicorn==19.3.0
hiredis==0.2.0
kombu==3.0.29
paramiko==1.16.0
psycopg2==2.6.1
PyChef==0.2.3
pycrypto==2.6.1
pytz==2015.7
PyYAML==3.11
redis==2.10.5
SQLAlchemy==1.0.9
wheel==0.24.0
